made the getmouse mechanism plus the measure distance command. modified the controller so that it had a signal emmitted when the mouse gets moved or clicked the scripting editr then hooks to both signals and we cache the mouse position in the hsfapi scripting enviroment and then evalute the script code this way the scripting enviroment will get the current mouse position.
made sure the vector function checked for distances of the input points, it was crashing, and then you would not know the error was related to this, accroding to visual studio it was showing the error on the viewport movemouse.
this is confusing, why doesnt it actually show the error where it happens.

added a get isocurve function to the hsf factory and used it in the userais shape to draw the isocurves on it.